up

Created-by: msprofDeveloper

Match-id-b27f2d52ab8e9d1d00748d7d425eb3a8f67c313b

+++ collector/dvvp/adx/stub/adx_prof_api.cpp
@@ -32,7 +32,7 @@ int DoAdxIdeCreatePacket(CmdClassT type, IdeString value, uint32_t valueLen, Ide
     IdeTlvReq req = (IdeTlvReq)sendBuf;
     req->type = type;
     req->dev_id = 0;
-    req->len = valueLen;
+    req->len = static_cast<int32_t>(valueLen);
     errno_t err = memcpy_s(req->value, mallocValueLen, value, valueLen);
     if (err != EOK) {
         MSPROF_LOGE("memory copy failed, err: %d", err);
